Moved plugin components to api folder and made API experimental
